app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.0
pydantic-settings==2.6.0
email-validator==2.2.0
python-multipart==0.0.9
